Input: cs40l26 - Allow STOP_PLAYBACK even if vibe_state is stopped
There is no need to prevent a STOP_PLAYBACK mailbox command
from the driver if vibe_state is 0 (VIBE_STATE_STOPPED).

Remove this check and replace it with a check for skip_delay.

If delay has been skipped for the particular RAM index
it is safe to skip the STOP_PLAYBACK command as well.

Change-Id: Ic0d315bb07c1ee7a0d90e76f648512fd7ccdef40
Signed-off-by: Fred Treven <fred.treven@cirrus.com>