Input: cs40l26 - Move HI-Z configuration to dsp_config()
The driver configures the POWER_ON_SEQUENCE to
set speaker output to HI-Z when in standby.

This was initially set in the wseq_init() function
but that is not the most appropriate place for
this configuration.

Move the code to dsp_config() which is a better context
for this to be done in.

Change-Id: If6cdb2a9b766ac4b751c1b7a9d32aa5ff13c03d3
Signed-off-by: Fred Treven <fred.treven@cirrus.com>