firmware: cirrus: Change error from -ENXIO to -EINVAL
Error type -ENXIO should only be used when there is no such
device or address. In cl_dsp_get_reg() and cl_dsp_get_flags()
this is not quite accurate.

The functions should use the error returned by
cl_dsp_get_coeff() if there is one and, if there is not,
return -EINVAL since the most likely issue would be an
invalid argument to reach that error case.

Change-Id: I94ac056c6db00fc603c11ed14d2b5044ea5f772d
Signed-off-by: Fred Treven <fred.treven@cirrus.com>