Input: cs40l26 - Allow for multiple debugFS instances
In case of more than one CS40L26 device,
debugFS directory mounting was failing the second
time due to there already being a directory with
the same name in the root debugFS directory.

This led to a kernel panic when an undefined
struct instance was freed by the driver.

Remove the kfree() invocation since debugFS
recursive remove procedure will take care of
this upon unloading the driver.

Create the directories with names using the node
name from the devicetree (e.g. "cs40l26a") and the
I2C bus number with the device's 7-bit I2C address.

For example: /sys/kernel/debug/cs40l26a-2-0043
indicates that the device is on I2C bus 2 and has
7-bit address 0x43.

Change-Id: Ia43243ba5820838c63152a24057960aea429577b
Signed-off-by: Fred Treven <fred.treven@cirrus.com>