Input: cs40l26 - Prevent hibernation before swapping firmware
The device could hibernate at a time when both PM runtime
and IRQs are disabled. This could cause unknown behavior
during the firmware swapping procedure which could lead
to corrupted data and an unresponsive DSP.

Send a PREVENT_HIBERNATE mailbox command so that
the device cannot hibernate during the swapping
procedure and the DSP remains in a known good
state.

Change-Id: I0bf039ccb5420eb24ce53d2153d75e0c082902b3
Signed-off-by: Fred Treven <fred.treven@cirrus.com>