Input: cs40l26 - Add support for Rich Trigger Haptics
Add sysfs control "rth_latch" that allows the user to
apply settings from RTH controls in the firmware.

The RTH settings can be programmed via the generalized
firmware sysfs interface and enforced on haptics
playback via the "rth_latch" control.

This feature is only supported on 0xA1 and 0xB1 devices.

Change-Id: I557b35b3edd627bfb98bff1c5b5f58cb83cceb3b
Signed-off-by: Fred Treven <fred.treven@cirrus.com>