Input: cs40l26 - Restore PLL properly if open-loop
set_pll_loop() checks if the driver needs to lock the REFCLK
input PLL (closed-loop)  or to put it in free-running
(open-loop) mode.

The function was checking to see if the input parameter
(pll_loop) was equal to 0 or 1. However, the input
was only masked when set_pll_loop() is called from
vibe_state_update() at the end of an ASP event; it was
not shifted to be properly compared directly to 1.

Use FIELD_GET to mask and shift bit 11 of the REFCLK_INPUT
value so that it is not rejected in set_pll_loop() if the
PLL needs to be put in open-loop mode.

Change-Id: I186d129c1f2126dd323505bc9a6eb44516cadd88
Signed-off-by: Fred Treven <fred.treven@cirrus.com>