Input: cs40l26 - Replace some instances of regmap_update_bits
It's better to use regmap_set_bits() or regmap_clear_bits()
over regmap_update_bits() if all bits in the bitfield are
1 (set_bits) or if they are all 0 (clear_bits), and if
the setting is always the same (e.g. it does not depend on
input or device configuration). This makes the code a bit easier
to read, clears up some space, and removes the need for certain
shift macros.

Replace regmap_update_bits() with regmap_clear_bits() and
regmap_set_bits() where appropriate.

Change-Id: I67521e631fd0fe5a163a5f75975e3e4df622db50
Signed-off-by: Fred Treven <fred.treven@cirrus.com>