Input: cs40l26 - Support I2C broadcast
Support the I2C broadcast feature by means of a new
device tree property "cirrus,i2c-broadcast-addr". When this
property is included, the device will allow writes to its
broadcast address. The first device in the device tree is
designated as the master, and any trigger start or stops sent
will be broadcasted.

In this configuration, hibernation is disabled.

Change-Id: I0b939c02597fe6152fd1b981709a1b08b3aaf64c
Signed-off-by: James Ogletree <jogletre@opensource.cirrus.com>