Input: cs40l26 - Get and enable regulators simultaneously
Use devm_regulator_bulk_get_enable() which will
instantiate and enable both regulators and allow
the resources to be device managed such that
the put and disable are handled by devres.

Remove cs40l26_supplies[] and explicit
calls to disabling the regulators in remove().

This ensures that there is no risk for regulator
enable/disable unbalance or accessing a regulator
resource that has not been instantiated.

Change-Id: I933a0be1ddf7968ddba223358b5ee10bd4914efa
Signed-off-by: Fred Treven <fred.treven@cirrus.com>