Input: cs40l26 - Default f0, ReDC values persist across firmware swap
Apply default f0 and ReDC values from devicetree properties
"cirrus,f0-default" and "cirrus,redc-default" respectively
any time firmware is loaded.

This will allow the values to persist or be re-applied when
firmware is swapped and if firmware loading is initially deferred.

It is left up to the user to overwrite the controls with calibrated
values if they do not want to use the default.

If "cirrus,f0-default" or "cirrus,redc-default" are not populated,
the default values from the firmware will be used.

Change-Id: I80bb26fe3127cc24a4edf7f6cd2bc1d07c955d60
Signed-off-by: Fred Treven <fred.treven@cirrus.com>