Input: cs40l26 - Explicitly assign f0 and ReDC default values
Guarantee that f0 and ReDC are not configured if the
respective devicetree properties "cirrus,f0-default"
and "cirrus,redc-default" aren't present by assigning
the values explicitly to 0 such that they are out
of the acceptable range. This will ensure that the
configurations are not performed and the device will
fall back to the default values in the firmware.

Prior to this change, it was techinically possible
for the variables to be assigned to unintended values
that are within the allowable range for f0 and ReDC.
This would cause the driver to configure the device
based on erroneous values. Assigning the variables
explicitly ensures that this situation cannot arise.

Change-Id: I46f05ab5082ecc9c5780b33997ddc3f7c01d2aee
Signed-off-by: Fred Treven <fred.treven@cirrus.com>