Input: cs40l26 - Add logic to report BUZZGEN playback to logging
Previously it was not possible to distinguish between a BUZZGEN
effect and a RAM wavetable effect if the index was > 128.

If the ROM flag is set in the DSP to host mailbox message and
the trigger index is larger than the number of ROM waveforms on
the device, the driver will determine the playback is a BUZZGEN effect.

For this logic to hold for GPIO triggers, the ROM/RAM bit
in the event map must be configured as ROM when a BUZZGEN
effect is mapped to a GPIO event.

Change-Id: I64f8d3bc53aecdbd720ea452c22aefd44a2e1916
Signed-off-by: Fred Treven <ftreven@opensource.cirrus.com>