ASoC: cs40l26: Reinitialize A2H after setting A2H Delay
Application of the "A2H Delay" mixer control which maps
directly to firmware's LRADELAYSAMPS requires the driver
to send A2H_REINIT in order to take effect.

Invoke the A2H_REINIT mailbox message so that the proper
A2H delay will be applied when the next stream is started.

Change-Id: Iad57b13c5492b8d841510788eb5142f09d3b4034
Signed-off-by: Fred Treven <ftreven@opensource.cirrus.com>