Input: cs40l26 - Log driver errors
Add sysfs control error_log which allows users to have access
to critical errors encountered and logged by the driver since
the last time the log was cleared.

The error_log control can be cleared by writing any
positive integer to it.

The driver catalogs the errors in a linked list and specifies
the type of error along with its Linux error code and
the function in which it occurred.

Change-Id: Id5d9642dd13eeeda07ac9ba8a3711156c3cff1ed
Signed-off-by: Fred Treven <ftreven@opensource.cirrus.com>