ASoC: cs40l26: Change A2H Delay implementation
The driver was limiting A2H Delay to 400 ms, but
the firmware control LRADELAYSAMPS actually has an
upper limit of 500 ms. Increase the limit to match this.

The description of the mixer control enforces this limit so
the logic in the a2h_delay_put() function that sets the control to the
maximum or minimum if the value set by the user is out of range actually
has no effect. Remove this logic.

Change the #define name to specify that the value is in ms.

Change-Id: I43ea34406e433d4e81953b4c74b784293c058baa
Signed-off-by: Fred Treven <ftreven@opensource.cirrus.com>