ASoC: SDCA: Add SDCA FDL data parsing
Add parsing of ACPI DisCo information specific to FDL (File DownLoad).
DisCo contains a list of File Sets which can be requested by the device
and within each of those a list of individual files to be downloaded to
the device. Optionally the contents of the files may also be present in
a special ACPI table, called SWFT (SoundWire File Table).

Change-Id: Idcccc80e7be3e3879c8daceefcc9b58b92323f3c
Signed-off-by: Maciej Strozek <mstrozek@opensource.cirrus.com>