Input: cs40l26 - Use GFP_ATOMIC when allocating memory for error logging
Using the GFP_KERNEL flag when issuing devm_kzalloc() allows sleep
during memory allocation which is not allowed under certain lock
situations which may occur when the driver is allocating structs
to add to the error logging linked list.

Avoid preemption errors in situations in which sleeping is not allowed
by using the GFP_ATOMIC flag instead.

Change-Id: I7b6e9542d77dbee31a4ca79afd6327c1cd57757a
Signed-off-by: Fred Treven <ftreven@opensource.cirrus.com>