Input: cs40l26 - Add separate control to check LS Calibration status
Runtime firmware no longer contains the firmware control
STATE_CAL_RETURN_CODE which was being used by the driver
to verify the validity of LS Calibration results before
providing them to the user via "ls_calibration_results".

Since it was removed, reading the results with runtime
firmware loaded would always fail.

Add a new sysfs control, "ls_calibration_status", which
can be used to check the status of LS Calibration after
running the procedure. This will allow the user to check
the results even if runtime firmware is loaded, and then
check the status to verify the validity of the results.

Change-Id: I2ca5d946d48fe9e1a9944fb69d0ab74137d3b734
Signed-off-by: Fred Treven <ftreven@opensource.cirrus.com>