Input: cs40l26 - Updating snprintf to sysfs_emit
In all cases where maxlen passed to snprintf is PAGE_SIZE, update
snprintf to sysfs_emit, where PAGE_SIZE is implicit, and remove
PAGE_SIZE as a passed argument.

Note that this is not all uses of snprintf in cs40l26-sysfs.c. There are
two uses of snprintf with maxlen other than PAGE_SIZE, and those are
left as snprintf.

Change-Id: Icfc5f7ff910d0c819cd409c1601e7405c08380d6
Signed-off-by: Liam Ogletree <liam.ogletree@cirrus.com>