Input: cs40l26 - Refactoring error_log_show
The prior implementation allocated a char buffer and used a combination
of successive snprintf and strncat calls to populate the allocated char
buffer before writing it to the passed argument buf.

This updated implementation uses sysfs_emit_at to remove the char buffer
allocation and successive snprintf and strncat calls. This speeds up the
function and eliminates the need for several temporary variables. Note
that 'count' is no longer needed because 'at' now tracks the amount of
bytes written to buf. Additionally, there is now no free_exit goto label.

We keep a temporary char buffer to snprintf the contents of the next
error entry to make sure we do not cut off an error entry due to
reaching the PAGE_SIZE limit.

Behaviorally, if there is an error calling sysfs_emit_at while
cycling through the elements in the error log, error_log_show will still
have written some number of elements to buf before exiting. In the prior
implementation, error_log_show would have written nothing to buf.
Furthermore, the prior implementation printed bytes up to PAGE_SIZE,
even if that caused error entries to be cut off at the end. This
implementation adds a check to only write whole error entries,
truncating if necessary. If the error log is truncated, we report this
using dev_info.

Change-Id: I3f831049702c7eabd6a66ac84b20cd8f6dfa7a0e
Signed-off-by: Liam Ogletree <liam.ogletree@cirrus.com>