Input: cs40l26 - Fix memory leak in cs40l26_handle_svc_le_nodes
If we fail the devm_kzalloc call to allocate memory (of size struct
cs40l26_svc_le) for any node, then we jump to the goto err label and
free all nodes previously allocated before freeing the pointer to an
array of pointers we received at the top of the function from the
devm_kcalloc call.

Previously, we would free the array of pointers to structs without
freeing any allocated structs, if present. Behaviorally, if node_count
is 0 (that is, if we fail on the first devm_kzalloc), then nothing
changes.

Change-Id: I5127980f2cf4337fede45b58c9746414a33ec4fa
Signed-off-by: Liam Ogletree <liam.ogletree@cirrus.com>