firmware: cirrus: Refactor reporting WMDR and WMFW information
In each of these two instances, we can print directly from the passed
'data' buffer instead of allocating memory and using memcpy to write the
'data' buffer to the allocated buffer. Doing so requires casting the u8
'data' buffer to be a char buffer. Additionally, we no longer need to use
'len'. We can reduce the code to only using the print statements and
remove the two handle_info_text functions.

Change-Id: Id6209106b3ddb48951fbed44751447b44dd6ef01
Signed-off-by: Liam Ogletree <liam.ogletree@cirrus.com>