Input: cs40l50 - Support I2C broadcast
Support the I2C broadcast feature by means of a new
device tree property "cirrus,i2c-broadcast-addr". When this
property is included, the device will allow writes to its
broadcast address. The first device in the device tree is
designated as the master, and any trigger start or stops sent
will be broadcasted.

Change-Id: I392b5cbb0512aaa4bee40ecccc31d92fb9fd31fb
Signed-off-by: David Rhodes <david.rhodes@cirrus.com>