Input: cs40l26 - Allow driver to finish probe after FW load fault
There are still aspects of the driver that are useful
even if RAM firmware loading fails. Prevent bailing
out of probe completely if there's a firmware loading failure
and conditionally set up the regmap chip IRQ with PM runtime
enabled or disabled depending on how far along the firmware
loading procedure went before the fault occurred.

This allows the user to check error_log to diagnose the
issue and to retry firmware loading via swap_firmware.

Change-Id: I258809c0ab753c4f66e3581995dedf9516c62ed0
Signed-off-by: Fred Treven <ftreven@opensource.cirrus.com>