ASoC: cs40l26: Suppress errors when A2H controls not present
When calibration firmware is loaded the A2H algorithm is
unavailable, and the LRADELAYSAMPS firmware controls has
been removed for A2H v5 (revision 10.0.0) as of firmware
10.0.27.

These are known scenarios so it can lead to confusion
when errors are thrown when accessing the "A2H Delay"
and "A2H Level" mixer controls. Suppress these errors
by checking the A2H algorithm is available for the loaded
firmware and that the A2H revision is earlier than
when A2H v5 was introduced. Instead, these mixer controls
will just have no operation.

Change-Id: I03ebd8db6fcc5cb84ee392f479d6a4448b3290ed
Signed-off-by: Fred Treven <ftreven@opensource.cirrus.com>