ASoC: SDCA: Add manual PM runtime gets to IRQ handlers
SDCA interrupt handling is a bit odd, the SDCA IRQ registers are
defined on a device level but the handling of the IRQ is at the
Function level. As such the regmap IRQ's PM runtime operations need to
be on the device itself to ensure those registers are available but an
additional runtime get is required for the Function child when the IRQ
is actually handled. Add the required manual PM runtime gets.

Change-Id: I07d1fa931dd6466a54e368914d1e05ae49159f69
Signed-off-by: Charles Keepax <ckeepax@opensource.cirrus.com>