Input: cs40l26 - Logging LS calibration version
The formatting of Q and RES values changed in FW version 10.0.32. This
patch reports the version of LS calibration used so customers can
properly interpret the outputs from LS calibration. LS calibration
version is appended to the output of the sysfs control
ls_calibration_results and logged when ls_calibration_results is written.

Change-Id: Id0c3f8f8816a768c08ac1926552d1881b7f3301c
Signed-off-by: Liam Ogletree <liam.ogletree@cirrus.com>