ASoC: SDCA: Update externally_requested flag to cover all requests
Currently there is a flag to indicate if an IRQ has been requested by
something outside the SDCA core, such that the core can skip requesting
that IRQ. However, it is simpler and more useful to always store the
allocated IRQ number. This will allow the core to see if the IRQ has
been requested, to perform additional operations on the IRQ, and
request IRQs in multiple phases.

Change-Id: Ibe8ec615a009df85c324246d5e12c9362b1d5c83
Signed-off-by: Charles Keepax <ckeepax@opensource.cirrus.com>