ASoC: SDCA: Rely less on the ASoC component in IRQ handling
In the future some IRQs (mostly the UMPs used during File
DownLoad) will need to run after the device has enumerated on the
bus but before the soundcard is actually constructed. As such
refactor more of the IRQ handling to use raw device and regmap
pointers, rather than accessing things through the component.

Change-Id: Icb65761aba50ca8f7963fca26f0914866904d18c
Signed-off-by: Charles Keepax <ckeepax@opensource.cirrus.com>