ASoC: SDCA: Rename SoundWire struct device variables
The SDCA core processes two different struct device's at various times,
usually it is processing the struct device associated with an individual
SDCA Function driver. However, there are times that it is processing the
struct device associated with the actual SoundWire peripheral, usually
the parent of the Function device.

To ensure this distinction remains clear in the code, rename the variable
when handling the actual SoundWire device to sdev.

Change-Id: I653d973d54e2b44fb8dd6aae37e04c71b092cde5
Signed-off-by: Charles Keepax <ckeepax@opensource.cirrus.com>