ASoC: SDCA: Correct FDL locking in sdca_fdl_process()
The current locking in sdca_fdl_process() locks over
sdca_ump_cancel_timeout() and the timeout work function takes the same
lock, this can lead to a deadlock if the work runs as part of the
cancel. To fix this move the guard() call to below the cancel timeout,
putting it outside the lock.

Fixes: e92e25f77748 ("ASoC: SDCA: Add UMP timeout handling for FDL")

Change-Id: I7401730ff30804e5413e80b3d0530a639b2f2059
Signed-off-by: Charles Keepax <ckeepax@opensource.cirrus.com>