Input: cs40l26 - error_log can be cleared on read or write
Adds sysfs control error_log_clear_method which allows the
user to configure the way in which error_log is cleared.

If error_log_clear_method is set to 0 (default), the log will
be cleared when error_log is read from. If error_log_clear_method
is set to 1, the log will be cleared by writing 1 to error_log
(legacy behavior).

Change-Id: I2d5683c6eda98846973d0d95f9e3de1af1b89627
Signed-off-by: Fred Treven <ftreven@opensource.cirrus.com>