ASoC: SDCA: Factor out jack handling into new c file
The jack code is perhaps a bit large for being in the interrupt
code directly. Improve the encapsulation by factoring out the
jack handling code into a new c file, as is already done for HID
and FDL. Whilst doing so also add a jack_state structure to hold
the jack state for improved expandability in the future.

Change-Id: I907eb43e4c41f5759f10e189483ca163da207e8d
Signed-off-by: Charles Keepax <ckeepax@opensource.cirrus.com>