Input: cs40l26 - Do not allow index 0 in OWT composite if not delay-only
The firmware explicitly treats index 0 in composite waveforms as a
"delay-only" section meaning that there will be no playback as it
is instead essentially a placeholder to allow a section that only
introduces additional delay.

Ensure that the composite is formatted properly by disallowing
any effect using index 0 that does not contain an associated delay
or does include an unnecessary duration (since the driver assumes
there will be no duration for index 0).

Change-Id: I2ffda32226e61c58941583c25dc8d874ce42293e
Signed-off-by: Fred Treven <ftreven@opensource.cirrus.com>