ASoC: cs40l26: Use sleep instead of completion for ASP SVC init
Asynchronous activity is blocked during DAPM events so the
interrupt processing required to mark the ASP SVC initialization
completion struct as "done" was never reached and the instance
of wait_for_completion_timeout() was always timing out.

Since we can accurately predict the SVC initialization time,
switch to using usleep_range() with a value calculated in
the core driver instead of the completion.

Change-Id: I2e9b1d878f9c54002249e9fff533bb4d512f6cb2
Signed-off-by: Fred Treven <ftreven@opensource.cirrus.com>