Input: cs40l26 - Use variable names in sizeof() as opposed to types
The preferred Linux coding standard for using sizeof() is to use
the name of the variable as opposed to its type where possible.

This avoids type mismatches in case the variable type changes,
improves maintainability, and ensures correctness for more
complex types.

Replace any instances using the variable type with the variable name.

Change-Id: I8e039af04775c882eb8bee607787a7c12956ae46
Signed-off-by: Fred Treven <ftreven@opensource.cirrus.com>