ASoC: SDCA: Device boot into the system suspend process
When system suspending the device may be powered off, this means all
state will be lost and the firmware may need to be re-downloaded. Add
the necessary calls to bring the device back up. This also requires that
that the FDL (firmware download) IRQ handler is modified to allow it to
run before runtime PM has been fully restored.

Change-Id: Ia2c55675e4fb37e69a38fb7eb0ea484ef7b8adb4
Signed-off-by: Charles Keepax <ckeepax@opensource.cirrus.com>