Input: cs40l26 - Track SVC initialization completion for streaming
Add boolean to check whether or not SVC initialization has been
completed by the DSP. Receiving an I2S trigger message from
the DSP-to-Host mailbox informs the driver that the high-frequency
pilot tone initialization for SVC streaming has completed.

This will be used in the codec driver to send a warning if delay
time was not sufficient for the initialization phase to finish.

Change-Id: I6916ad9e252b35e84b773e81f47f01237718da3e
Signed-off-by: Fred Treven <ftreven@opensource.cirrus.com>