ASoC: cs40l26: Only swap to BCLK early if using ASP Dout
If ASP Dout is enabled for signal recording, REFCLK_INPUT
needs to be swapped to using BCLK in component_set_sysclk()
since this callback is invoked prior to DAPM powering up
for ASP streaming.

Doing this prevents the device from receiving interrupts
until the stream has been fully powered up. This cannot
be allowed in cases for which ASP Dout is not configured.

Change-Id: Id72dcf24200f79168cd4a2d5cf634a03036dfd2d
Signed-off-by: Fred Treven <ftreven@opensource.cirrus.com>