ASoC: SDCA: Change class driver to depend on SND_SOC_SDCA
Currently the class driver selects SND_SOC_SDCA however as select
ignores dependencies this can lead to situations where the SDCA support
is built without ACPI which is currently not supported. Change the setup
such that the class driver depends on SND_SOC_SDCA to ensure the
dependencies are met.

Change-Id: I52006f51170ff4959771a71970532e3414e983d8
Signed-off-by: Charles Keepax <ckeepax@opensource.cirrus.com>