ASoC: SDCA: Change API for sdca_irq_enable()
Update the API for scda_irq_enable() to have an _early version of
the function rather than passing in a flag. The flag was confusing when
that API was called since it is just true/false it looked like it was
enabling/disabling the IRQ rather than doing early/late.

Change-Id: Id142c38e6de93e338138aa6ec3e11873f38f37ab
Signed-off-by: Charles Keepax <ckeepax@opensource.cirrus.com>