Input: cs40l26 - Ensure DBC is disabled when calibration firmware loaded
If runtime firmware was loaded along with a Dynamic Boost Control (DBC)
tuning file before swapping to calibration firmware, the driver was
mistakenly enabling DBC which is not officially supported in
calibration firmware.

Explicilty set the dbc_tuning_loaded boolean to false at the start
of the firmware load procedure so the driver does not enable it
for calibration.

Change-Id: I8ec8d2bd93f42e3d483e82a0d84c672d98069ee6
Signed-off-by: Fred Treven <ftreven@opensource.cirrus.com>